% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations2nc.R
\name{stations2nc}
\alias{stations2nc}
\title{Station data 2 netCDF export}
\usage{
stations2nc(
  data,
  NetCDFOutFile = "out.nc4",
  missval = 1e+20,
  globalAttributes = NULL,
  varAttributes = NULL,
  prec = "float",
  compression = 4,
  shuffle = FALSE,
  standardName = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data}{A grid data object coming from \code{\link[loadeR]{loadGridData}} or \code{\link[transformeR]{interpGrid}} 
or the function \code{\link[loadeR.ECOMS]{loadECOMS}} of package \pkg{loadeR.ECOMS}.}

\item{NetCDFOutFile}{Name of the file created by the function. (default to "out.nc4")}

\item{missval}{Missing value codification (default to 1e20)}

\item{globalAttributes}{Optional. A list of global attributes included in the NetCDF file. Same format as \code{varAttributes}.}

\item{varAttributes}{Optional. List of attributes to be included in the variable written in the NetCDF file. (NULL, default).
It has the format \code{list("name_of_attribute1" = attribute1, "name_of_attribute2" = attribute2} etc.}

\item{prec}{Precision to write the attribute. If not specified, the written precision is the same as the variable whose 
attribute this is. This can be overridden by specifying this argument with a value
 of \code{"short"}, \code{"float"}, \code{"double"}, or \code{"text"}.}

\item{compression}{If set to an integer between 1 (less compression) and 9 (highest compression), enables 
compression for the variable as it is written to the file. Turning compression on forces the 
output file to netcdf v4 format (the default), not compatible with older software 
that can only handle version 3 files.}

\item{shuffle}{Logical. Turns on (\code{TRUE}) or off (\code{FALSE}, the default) the shuffle filter.
According to netcdf documentation, turning the shuffle filter on can improve compression for integer variables.
 Turning the shuffle filter on forces the output file to netcdf v4 format (the default), 
 not compatible with older software that can only handle version 3 files.}

\item{standardName}{Standard name of the variable.}

\item{verbose}{Optional. If set to \code{TRUE}, switches \code{\link[ncdf4]{nc_create}} to verbose mode}
}
\value{
A NetCDF-4 file with the variable and attributes defined in the inputs.
}
\description{
Export a loadeR grid to NetCDF
}
\examples{
\dontrun{
require(transformeR)
data("VALUE_Iberia_tas")
# Name of output file:
fileName <- "tas_VALUE_DJF_Iberia.nc4"
# Including a global attribute:
globalAttributeList <- list("institution" = "Santander MetGroup, http://www.meteo.unican.es/")
# Including two variable attributes:
varAttributeList <- list(var_attr1 = "one_attribute", var_attr2 = "another_attribute")
# Create file:
stations2nc(data = VALUE_Iberia_tas,
        NetCDFOutFile = fileName,
        missval = 1e20,
        prec = "float",
        globalAttributes = globalAttributeList,
        varAttributes = varAttributeList)
}
}
\references{
\itemize{
\item David Pierce \email{dpierce@ucsd.edu}, Interface to Unidata netCDF (version 4 or earlier) format data files, http://dwpierce.com/software
}
}
\author{
S. Herrera, W. Franssen, J. Bedia and M. Iturbide
}
