% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid2nc.R
\name{grid2nc}
\alias{grid2nc}
\title{Grid 2 netCDF export}
\usage{
grid2nc(
  data,
  NetCDFOutFile = "out.nc4",
  missval = 1e+20,
  globalAttributes = NULL,
  varAttributes = NULL,
  prec = NULL,
  compression = 4,
  shuffle = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{data}{A climate4R grid data object (\url{http://www.meteo.unican.es/climate4R})}

\item{NetCDFOutFile}{Name of the file created by the function. (default to \code{"out.nc4"})}

\item{missval}{Missing value codification (default to \code{1e20})}

\item{globalAttributes}{Optional. A list of global attributes included in the NetCDF file. Same format as \code{varAttributes}.}

\item{varAttributes}{Optional. List of attributes to be included in the variable written in the NetCDF file. Default to \code{NULL}.
It has the format \code{list("name_of_attribute1" = list(attribute1, prec of attribute1), "name_of_attribute2" = list(attribute1, prec of attribute1)} etc.}

\item{prec}{Precision to write the attribute. If not specified, the written precision is given by the variable's
corresponding attribute. This can be overridden by specifying the following argument values:
 \code{"short"}, \code{"float"}, \code{"double"}, or \code{"text"}.}

\item{compression}{Integer value between 1 (less compression) and 9 (highest compression), enabling 
compression for the variable as it is written to the file. Note that turning compression on forces the 
output file to netcdf v4 format (the default), not compatible with older software 
that can only handle version 3 files.}

\item{shuffle}{Logical. Turns on (\code{TRUE}) or off (\code{FALSE}, the default) the shuffle filter.
According to netcdf documentation, turning the shuffle filter on can improve compression for integer variables.
 Note that turning the shuffle filter on forces the output file to netcdf v4 format (the default), 
 not compatible with older software that can only handle version 3 files.}

\item{verbose}{Optional. If set to \code{TRUE}, switches \code{\link[ncdf4]{nc_create}} to verbose mode}
}
\value{
A NetCDF-4 file with the variable and attributes defined in the inputs.
}
\description{
Export a loadeR grid to NetCDF
}
\examples{
\dontrun{
library(loadeR.2nc)
data(tx) # A climate4R grid
# Name of output file:
fileName <- "/tmp/tasmax_WFDEI_JJA_W2001_2010.nc"
# Including a global attribute:
globalAttributeList <- list("title" = "Daily mean maximum temperature data from the WFDEI Database over Southern Europe used as an example of netCDF export from climate4R",
                            "institution" = "SantanderMetGroup, <http://www.meteo.unican.es>", 
                            "source" = "climate4R <https://github.com/SantanderMetGroup/climate4R>")
# Including variable attributes. The last one not really correct, but for illustration purpose:
varAttributeList <- list("name" = list("tasmax", prec = "text"),
                         "description" = list("Daily maximum near surface air temperature", prec = "text"),
                         "level" = list(2L, prec = "short"))
# Create file:
grid2nc(data = tx,
        NetCDFOutFile = fileName,
        missval = 1e20,
        prec = "float",
        globalAttributes = globalAttributeList,
        varAttributes = varAttributeList)
}
}
\references{
\itemize{
\item David Pierce \email{dpierce@ucsd.edu}, Interface to Unidata netCDF (version 4 or earlier) format data files, http://dwpierce.com/software
}
}
\author{
S. Herrera, W. Franssen and J. Bedia
}
