% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx.R
\docType{data}
\name{tx}
\alias{tx}
\title{Maximum temperature grid dataset}
\format{a \pkg{loadeR} grid}
\source{
Data from the WATCH Forcing Dataset ERA-Interim retrieved from the ECOMS User Data Gateway (http://meteo.unican.es/trac/wiki/udg/ecoms).
}
\usage{
data(tx)
}
\description{
A Grid containing data of summer maximum 2m temperature for the Mediterranean (2001-2010)
}
\references{
Weedon, G.P et al. 2011. Creation of the WATCH Forcing Data and its Use to Assess Global and Regional Reference Crop Evaporation over Land during the Twentieth Century. Journal of Hydrometeorology 12, 823-848. doi:10.1175/2011JHM1369.1
}
\author{
Santander MetGroup (http://meteo.unican.es)
}
\keyword{data}
